+++ client/package.json
@@ -5,14 +5,15 @@
     "main": "client.ts",
     "license": "BSD-4-Clause",
     "scripts": {
-        "devA": "yarn tsx client.ts 0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d A 0 0",
-        "devB": "yarn tsx client.ts 0x5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a B 0 4",
-        "devC": "yarn tsx client.ts 0x7c852118294e51e653712a81e05800f419141751be58f605c371e15141b007a6 C 4 0"
+        "devA": "yarn tsx client.ts 59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d A 0 0",
+        "devB": "yarn tsx client.ts 5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a B 0 4",
+        "devC": "yarn tsx client.ts 7c852118294e51e653712a81e05800f419141751be58f605c371e15141b007a6 C 4 0"
     },
     "dependencies": {
         "dotenv": "^16.3.1",
         "ethers": "^5.7.2",
-        "socket.io-client": "^4.7.1"
+        "socket.io-client": "^4.7.1",
+        "viem": "^1.18.9"
     },
     "devDependencies": {
         "ts-node": "^10.9.1",
